Revise parsing of generic type argument lists.
We implemented parsing of generic type applications as part of an intern
project.  The parsing code relied on detecting an occurrence of generic
function (or a function with a generic bounds-safe interface in a checked
scope). If either condition was met, the code assumed that a generic type
argument list must follow.  It tried to parse that type argument list to
create a type application.

Reimplement parsing of generic type applications so that it does not depend on
information from semantic checking.  Relying on information from semantic
checking is a potential sign that that grammer for ambiguous.

We treat generic type applications as postfix expression, with the
grammar production:

postfix-expression:
   postfix-expression < type argument list>

where type argument list consists of a comma separated list of zero
more type arguments.

This isn't ambiguous because the `<` operator cannot currently be
followed by a type name or '>'.  We just need to check the first token after
'<' to see if it is a type name or '>'.  Of course, there is an ambiguity in
the C grammar that an identifier could either by the name of a type or
the name of a variable.  We haven't made this worse.

Add a new diagnostic message for the case where a non-generic function
is applied to type arguments.  Before, there was just a parsing failure
of the form "expected expression".

This change partly addresses Github issue #559 (address of operator
on function with generic type parameters).  It now allows the address-of
operator to be applied to generic functions with type arguments.

Testing:
- Passes existing tests, except for these where a generic function
  is called without applying it to type arguments.