Improve transformation of positional parameters.
Change the transformation of positional parameters in TreeTransform.h
to use the transformed types for parameters.  Do this by using the
types passed into TransformExtendedParameterInfo.

Before this, the code was just transforming the type for the positional
parameter based on however types were being transformed.,  This worked fine
for a type-only transform, but didn't work properly when the parameter type
was transformed based on some other informaton in the program (such as a
bounds-safe interface).

This fixes issue #484.

Testing:
- Add a regression test based on the case described in issue #484.