Fix for compiler assert.
This change fixes the compiler assert reported in Github issue #575. We are
inserting temporary variables for strings so that we can describe their
bounds.  The assert was checking that we don't compute a temporary
variable more than once in an expression.

The problem is that we are copying the expression that creates and binds the
temporary into the bounds.   The bounds is used at runtime in the
dynamic_bounds_cast.  This causes the temporary to computed twice, which is
what the assert is protecting against.  Since we've already computed the value
into a temporary, the fix is to replace the binding with a read of the
temporary.

Testing:
- Added simple repro case that caused a compiler crash.
- Passed local testing and repro case.  Still needs a runtime test that the
  right thing is happening.