Implicit void pointer casts for conditional expressions in checked scopes.
Disallow some implicit void pointer casts for arms of conditional expressions.
In C, if one arm has a void pointer type and the other has a pointer to a
non-void type, the expression with the pointer to a non-void type is cast
implicitly to have a void pointer type.

In a checked memory-safe scope, when the non-void type contains checked
pointer, we do not allow the cast.   The void pointer could be cast to
some other pointer type that could be used to corrupt the checked pointer.
We do allow casts between pointers to void and pointers to data not
containing checked pointers in checked memory-safe scopes.