Remove space in pattern matching for function beginnings in LLVM IR.
The text representation of a function in LLVM IR may or may not have a modifier
'dso_local' after the define keyword, depending on the target plaform.  For
example:
  define dso_local i8* ptrGenericTest
We used the pattern 'define {{.*}} i8*' to match this. Apparently this causes
problems for FileCheck's pattern matcher, with some sort of white space
mismatch when 'dso_local' is omitted. Use the pattern 'define {{.*}}i8*'
(which does not have sapce after the }}) instead.