Fix compiler assertion when subscripting a call expression.
We introduce temporary variables to hold the results of calls
to functions that return pointers with bounds.  This lets us
express the bounds of the result of the call in the IR.  There is
an assertion failure during code generation.  The wrong map is
being used to look up the LLVM value for the temporary variable
for an rvalue.  Use the right map.

Testing:
- Passes tests of subscripting call expressions.