Add zero versus unsigned expression comparison (#599)
Everytime in bounds declaration checking we encounter a comparison
between zero and an unsigned integer, we can use the fact that
0 <= any unsigned integer expression.

In this commit, the representation of ranges has been extended. Instead
of ConstantSizedRange, a new representation is introduced: BaseRange.

InRange() checks whether the range is constant-sized or
variable-sized, and compare lower/upper offsets based on that.

CreateBaseRange() is responsible for setting the correct RangeType to a
range.