Extend type substitution to handle instantiated records
When doing a substitution T1 [U -> T2] (substitute type T2 for type
variable U in T1), we need to handle the case where T1 is a type
application whose arguments potentially need to be rebound.

e.g.

```
struct Box<T> {};
struct List<T> {
  struct Box<T> box;
};
struct List<int> l;
```

In the snippet above, while generating the RecordDecl for
'l', we encounter the field 'struct Box<T> box', which is
an instantiated RecordType, and where we need to rebind the type
variable T.

Contrast with a hypothetical field 'struct Box<int> box', where the
argument to 'Box' is fixed and so won't change.

Here's how we transform an instantiated RecordType:
1) for each type argument to the the RecordType, perform the subtitution
1) for each type argument A of the record type, recursively perform
the substitution A [U -> T2].
2) collect the new set of (substituted) type arguments
3) re-instantiate the base generic type ('Box' for the example above)
with the new set of arguments