Added the _MMSafe_ptr pointer in the Clang AST.
Added a new type of checked pointer  _MMSafe_ptr (also used as
mmsafe_ptr for brevity). This new type of pointer can only point
to a struct object, and all the restrictions imposed on the _Ptr
type of pointer apply to it. It is meant to provide safe memory
management: dereferencing pointers of this kind is guaranteed to
be free from Use-after-Free errors.

- Added this new pointer type in the CheckedPointerKind enum class in
  include/clang/AST/Type.h.
- Added a isCheckedPointerMMSafeType() function to
  include/clang/AST/Type.h.
- Added this new pointer type in declaration and type specifier related
  source files including include/clang/Basic/Specifiers.h,
  include/clang/DeclSpec.h, lib/Parse/ParseDecl.cpp,
  lib/Sema/DeclSpec.cpp, lib/Sema/SemaType.cpp (this file als contains
  the code that builds _MMSafe_ptr as a PointerType.).
- Added _MMSafe_ptr as a keyword in include/clang/Basic/TokenKinds.def.

Also wrote tests to test parsing this new type of pointer in
tests/parsing/pointer_types.h in the checkedc repository.
The current changes of this repo and the checkedc repo passed all
the checkedc and clang tests.