Change the logic for when we delay type applications
Before this change, we use to suspend all type applications that
happened during a field creation.

This is in fact incorrect, because some applications might refer
to definitions that have already been completed, and so shouldn't
be suspended.

The new approach is only suspend applications that can't be completed
because the base (the underlying generic struct) isn't complete.