Added check on the type of the pointee of a _MMSafe_ptr pointer.
Added check on the type of the pointee of a _MMSafe_ptr pointer.
Only struct types are allowed to be pointed by a _MMSafe_ptr pointer.

Also renamed "MMSafe_ptr" to "MMSafe" in the CheckedPointerKind
enum class in include/clang/AST/Type.h. This is for consistency:
Neither of _Array_ptr and _Nt_array_ptr has the "_ptr" suffix in
this enum class.

Also added corresponding tests in the checkedc repo. The changes
induced in this commit passed the checkedc and clang tests.