Improve equivalence checking of pointer arithmetic expressions (#596)
A new equivalence checking function has been added in this change
(EqualPointerArithOffset). This function is used for comparing two
bounds (i.e., Base + Offset) where one bound is following pointer
arithmetic form and the other one is in bytewise arithmetic form.

Note that this commit assumes that overflow checking is not a concern.