Fixed a bug in checking the pointee type of a _MMSafe_ptr<T>.
_MMSafe_ptr<T> is only allowed to point to a struct type object.
However, one special case is when _MMSafe_ptr<T> is used in a generic
function. In a generic function, a generic type should be allowed to
be the type of the pointee of a _MMSafe_ptr<T>.

Also added test code to test using _MMSafe_ptr<T> in a generic function.
The commit of the test is at https://github.com/microsoft/checkedc/commit/b4f90d9285e194f558ba1ec85bc27c62d11c4853