Fixed incorrect return and arugment types for _MMSafe_ptr on X86_64.
The generated LLVM IR argument and return types for _MMSafe_ptr were
incorrect. For example, for the following code,

    mmsafe_ptr<struct node> foo(void) {
               mmsafe_ptr<struct node> p;
               return p;
    }

the generated LLVM IR code is

    define dso_local %struct.node* @foo() #0 {
    entry:
      %retval = alloca { %struct.node*, i64 }, align 8
      %p = alloca { %struct.node*, i64 }, align 8
      %0 = load { %struct.node*, i64 }, { %struct.node*, i64 }* %p, align 8
      store { %struct.node*, i64 } %0, { %struct.node*, i64 }* %retval, align 8
      %coerce.dive = getelementptr inbounds { %struct.node*, i64 }, { %struct.node*, i64 }* %retval, i32 0, i32 0
      %1 = load %struct.node*, %struct.node** %coerce.dive, align 8
      ret %struct.node* %1
    }

The return type should be a struct containing a pointer to struct node
and a 64-bit integer as the ID for the struct node; however it was
coerced to a single pointer. For X86_64 ABI, during the LLVM IR
code generation, the compiler calls a function X86_64::classify() to
classify the the argument and return types of a function, and it
treats a pointer type as an integer. In this commit, we added check on
whether the pointer is a _MMSafe_ptr and initialize the type properly.

Currently we have only fixed the the _MMSafe_ptr<T> generation for
function arguments and return values on X86_64. Further investigation
on 32-bit platforms and Windows-specific ABI is needed.