Uniform the _MMSafe_ptr<T> return types of a generic function.
If a generic function returns a _MMSafe_ptr<T>, there would be a type
mismatch between the generic return type and the real return type.
For example, for the following code:

    // callar code:
    foo<struct Data>();

    // callee code:
    mmsafe_ptr<T> p;
    .......
    return p;

the generated IR code of the return type for the generic function is
{i8*, i64} while the IR code from the caller's view is
{struct.Data*, i64}. The compiler would try to do a bitcast from
the generic type to the real type. However, this bitcast would cause
an assertion failure because bitcast between struct type is
disallowed. For this, situation, before the bitcast attempt, we added
code to mutate the first field of the generic return type to be the
same as the first field of concrete return type.

To successfully compile this commit, you need update the llvm repo.
The corresponding commit of the llvm repo is:
https://github.com/microsoft/checkedc-llvm/commit/960ae580fce7c1043bb3e2a9d66d946e92f9212f