Uniform the _MMSafe_ptr<T> argument type of a generic function.
If a generic function has a _MMSafe_ptr<T> paramter, there would be
a type mismatch between the generic type T and the concrete type
passed to this function when it is called. We implement _MMSafe_ptr
as a struct and LLVM flatterns the struct to two scalar arguments
(a pointer and a 64-bit integer of ID) when _MMSaf_ptr is used as a
parameter. It raises an assertion failure when the concrete pointer
type (a "struct.xxx*") does not match the generic type (a "i8*).

We added code to cocerce the concrete pointer type to be "i8*" during
the concrete argument list generation. Coercing the generic type to
the concrete type doesn't work beause it breaks the prototype of the
function and the compiler generates a "fatal error: error in backend:
Broken function found, compilation aborted!".