Add generic structs to Checked C
This PR adds generic structs to Checked C.

Generic structs are an extension to the Checked C type system that allows us to define structs with type variables that are quantified over an arbitrary type. The goal of generic structs is to eliminate cases where we'd previously need void pointers, because of C's lack of polymorphism.

Example: generic lists

To define a list that can contain pointers of an arbitrary type (arbitrary but consistent across list elements), we'd previously use void pointers:
```C
struct List {
  void *head;
  struct List *tail;
};
```

With generic structs, we can avoid the void pointer
```C
struct List _For_any(T) {
  T *head;
  struct List<T> *next;
};
```

An instantiation of the List generic is then: `struct List<int> li;` or `struct List<char> lc;`. The fields then have the expected types.

# How is this implemented?

The changes span only the parser and semantic analysis. There are no changes to code generation.

## Parser Changes

We need to change two things in the parser:
1) We need to accept `_For_any()` clauses in struct definitions (the _For_any goes after the struct name, as the example above shows).
2) We need to parse type applications, as in `struct List<int>`.

## Semantic Analysis Changes

Both a generic struct and a type application (type application = an instantiation of a generic struct) are represented with `RecordTypes`. The RecordType now has fields/methods for all the operations below:
1) is this record a generic/type application
2) get me the type parameters/type arguments of a RecordDecl

The type substitution operation is re-used from the existing implementation of generic functions.
See CheckedCSubst.cpp for details.

## Expanding Cycles

A lot of the complexity in the PR comes from having to deal with problematic struct definitions that are "cyclic" in a technical sense defined in CheckedCSubst.cpp (see "expanding cycles").

For example, consider

```C
struct Foo _For_any(T) {
  struct Foo<struct Foo<T> > *next;
};
```

In order to instantiate `struct Foo<int>`, we need to perform the substitution `int/T` in the struct's body, requiring the following instantiations: `Foo<Foo<int>> => Foo<Foo<Foo<int>>> => ...`. 

This leads to an infinite number of type applications. We therefore need to rule out certain struct definitions like the one above. The test for invalid structs is called an "expanding cycles check". See CheckedCSubst for details, but the algorithm is adapter from previous work on the CLR for e.g. C# (it's also used in Java and Scala).

## Tests

There are clang-specific tests under `tests/Sema/CheckedC/generic-structs`. The CheckedC repo also has more comprehensive spec tests.