Improve equivalence checking of pointer arithmetic expressions (#643)
A new equivalence checking function `EqualExtended()` has been added in this change.
This function is used for comparing two
bounds after converting the Offsets into a standard form `Constant OP Variable` depending on the type of the Base. In other words, the standard form comprises of a constant part, a variable part, and a boolean indicating the sign of the multiplication OP.

Given `array_ptr<T> p`:
- For `(char *)p + e1` or `(unsigned char *)p + e1`, `Constant = 1`, `Variable = e1`.
- For `p + e1`, `ConstantPart = sizeof(T)`, `VariablePart = e1`.
In both the above cases, the sign of OP is determined by the size of e1.