Generate LLVM IR code for _MMSafe_ptr. (#653)
Summary of changes: 
- Modified the LLVM IR code generator to generate IR code for a  _MMSafe_ptr when it is 
  - declared in a function
  - an argument passed to a generic function
  - a return value from a generic function
- Added check on declaration of _MMSafe_ptr: it must be initialized. 
- Added _MMSafe_ptr support in the getType() function of clang::Address class. Now it returns the inner pointer type of a _MMSafe_ptr.

* Added code of building a _MMSafe_ptr type in LLVM IR.

In LLVM IR, a _MMSafe_ptr is essentially an aggregate type of object
that contains a real pointer to the pointee (a struct object) and
an integer ID. We chose to implement it as llvm::StructType.

- Currently we hardcoded the ID as a 64-bit integer. This gives us
  extremely highly randomness compared with a 32-bit integer.
  However, on a 32-bit platform, such as most of the systems running on
  IoT devices, using a 64-bit integer may affect the peformance of
  programs. Maybe we should use an "unsigned long" type for the ID.

- We implemented the function that generates a struct for a _MMSafe_ptr
  in llvm/lib/IR/Type.cpp, because the function (PointerType::get())
  that generates a non-_MMSafe_ptr pointer to its pointee for LLVM IR
  is in this file.  So, to correctly compile and run Clang with this
  new commit, the clang-llvm repo also needs to be updated.

- Currently we have only finished the code of generating a struct for a
  _MMSafe_ptr pointer, but have not finished connecting the pointer to
  its pointee.  For example, the current compiler can compile the
  following code correctly:

    mmsafe_ptr<struct node> *p;

  The compiler will generate the following IR code:

    %p = alloca { %struct.node*, i64 }, align 8

  Now we will get compilation error if we try to assign something
  to p.

* Added code of building a _MMSafe_ptr type in LLVM IR.

In LLVM IR, a _MMSafe_ptr is essentially an aggregate type of object
that contains a real pointer to the pointee (a struct object) and
an integer ID. We chose to implement it as llvm::StructType.

- Currently we hardcoded the ID as a 64-bit integer. This gives us
  extremely highly randomness compared with a 32-bit integer.
  However, on a 32-bit platform, such as most of the systems running on
  IoT devices, using a 64-bit integer may affect the peformance of
  programs. Maybe we should use an "unsigned long" type for the ID.

- We implemented the function that generates a struct for a _MMSafe_ptr
  in llvm/lib/IR/Type.cpp, because the function (PointerType::get())
  that generates a non-_MMSafe_ptr pointer to its pointee for LLVM IR
  is in this file.  So, to correctly compile and run Clang with this
  new commit, the clang-llvm repo also needs to be updated.

- Currently we have only finished the code of generating a struct for a
  _MMSafe_ptr pointer, but have not finished connecting the pointer to
  its pointee.  For example, the current compiler can compile the
  following code correctly:

    mmsafe_ptr<struct node> *p;

  The compiler will generate the following IR code:

    %p = alloca { %struct.node*, i64 }, align 8

  Now we will get compilation error if we try to assign something
  to p.

* Fixed incorrect return and arugment types for _MMSafe_ptr on X86_64.

The generated LLVM IR argument and return types for _MMSafe_ptr were
incorrect. For example, for the following code,

    mmsafe_ptr<struct node> foo(void) {
               mmsafe_ptr<struct node> p;
               return p;
    }

the generated LLVM IR code is

    define dso_local %struct.node* @foo() #0 {
    entry:
      %retval = alloca { %struct.node*, i64 }, align 8
      %p = alloca { %struct.node*, i64 }, align 8
      %0 = load { %struct.node*, i64 }, { %struct.node*, i64 }* %p, align 8
      store { %struct.node*, i64 } %0, { %struct.node*, i64 }* %retval, align 8
      %coerce.dive = getelementptr inbounds { %struct.node*, i64 }, { %struct.node*, i64 }* %retval, i32 0, i32 0
      %1 = load %struct.node*, %struct.node** %coerce.dive, align 8
      ret %struct.node* %1
    }

The return type should be a struct containing a pointer to struct node
and a 64-bit integer as the ID for the struct node; however it was
coerced to a single pointer. For X86_64 ABI, during the LLVM IR
code generation, the compiler calls a function X86_64::classify() to
classify the the argument and return types of a function, and it
treats a pointer type as an integer. In this commit, we added check on
whether the pointer is a _MMSafe_ptr and initialize the type properly.

Currently we have only fixed the the _MMSafe_ptr<T> generation for
function arguments and return values on X86_64. Further investigation
on 32-bit platforms and Windows-specific ABI is needed.

* Uniform the _MMSafe_ptr<T> return types of a generic function.

If a generic function returns a _MMSafe_ptr<T>, there would be a type
mismatch between the generic return type and the real return type.
For example, for the following code:

    // callar code:
    foo<struct Data>();

    // callee code:
    mmsafe_ptr<T> p;
    .......
    return p;

the generated IR code of the return type for the generic function is
{i8*, i64} while the IR code from the caller's view is
{struct.Data*, i64}. The compiler would try to do a bitcast from
the generic type to the real type. However, this bitcast would cause
an assertion failure because bitcast between struct type is
disallowed. For this, situation, before the bitcast attempt, we added
code to mutate the first field of the generic return type to be the
same as the first field of concrete return type.

To successfully compile this commit, you need update the llvm repo.
The corresponding commit of the llvm repo is:
https://github.com/microsoft/checkedc-llvm/commit/960ae580fce7c1043bb3e2a9d66d946e92f9212f

* Uniform the _MMSafe_ptr<T> argument type of a generic function.

If a generic function has a _MMSafe_ptr<T> paramter, there would be
a type mismatch between the generic type T and the concrete type
passed to this function when it is called. We implement _MMSafe_ptr
as a struct and LLVM flatterns the struct to two scalar arguments
(a pointer and a 64-bit integer of ID) when _MMSaf_ptr is used as a
parameter. It raises an assertion failure when the concrete pointer
type (a "struct.xxx*") does not match the generic type (a "i8*).

We added code to cocerce the concrete pointer type to be "i8*" during
the concrete argument list generation. Coercing the generic type to
the concrete type doesn't work beause it breaks the prototype of the
function and the compiler generates a "fatal error: error in backend:
Broken function found, compilation aborted!".

* Added check on unitialized _MMSafe_ptr<T>.

_MMSafe_ptr<T> must be initialized.

* Added _MMSafe_ptr support for getType() of clang::Address.

When the Address contains a _MMSafe_ptr, getType() method should
return the inner pointer of _MMSafe_ptr.