Added _MMSafe_ptr support in clang::Address.
When a new Address is created, check if the pointer is a _MMSafe_ptr,
and reset the type if it is. Also added a public method to check if
the Address contains a _MMSafe_ptr. Without this method an Address
would lose the information that it actually contains a _MMSafe_ptr
after the type resetting.