Add an AST identifier to PersistentSourceLocation.
In certain cases, the file/line/column information stored in the
PersistentSourceLocation class is not enough to uniquely disambiguate
definitions from one another.  To support these cases, this commit
adds a "reproducible" AST node identifier as a distinguishing feature.

According to the "Clang Front End for LLVM Developers' List":

    "If you want something that's more stable across runs (pointers
    would obviously change every time), you can have a look at the
    getID() method in various AST nodes, which produces an integer
    that doesn't change across runs but does change across host
    machines."

Which should work fine for our purposes.

This fixes #18.