Implemented the "*" operator for _MMSafe_ptr.
*Added _MMSafe_ptr support in clang::Address.

When a new Address is created, check if the pointer is a _MMSafe_ptr,
and reset the type if it is. Also added a public method to check if
the Address contains a _MMSafe_ptr. Without this method an Address
would lose the information that it actually contains a _MMSafe_ptr
after the type resetting.

*Revised how Address handles _MMSafe_ptr.

For the clang::Address class, added new private fields to back up
the original pointer type and the inner pointer type in case of
_MMSafe_ptr. Also added a method to reset the pointer type from the
_MMSafe_ptr type to its inner pointer type and a method to restore
the pointer type to be the original _MMSafe_ptr type.

*Implemented the "*" dereference operator for _MMSafe_ptr.

When a mmsafe_ptr is dereferenced by the "*" operator and after the
null pointer checking is done, the compiler inserts a
DynamicStructIDCheck() function to check if the ID of the mmsafe_ptr
matches the ID of the pointee (a struct).

The algorithm is straightforward:

1. Extract the ID from the mmsafe_ptr.
2. Extract the ID from the struct (always the first field)
3. Insert an integer comparison instruction (ICmpInst) to compare the
   two IDs.
4. Continue executing if the IDs match; otherwise abort the program.

We also implemented a customized memory allocator and a deallocator,
and added tests for this commit. The customized allocator/deallocator
and the tests are in the checkedc repo:
https://github.com/microsoft/checkedc/commit/e4e3036386b82d848f82091bebddbdaa5b46d8b7