Using comparison facts for proving more bounds declarations at compile time
This change adds a forward dataflow analysis for collecting comparison facts in every function for each CFG block. These facts are then used for proving bounds declarations at compile time.

 
The main function that performs the dataflow analysis is `AvailableFactsAnalysis::Analyze()`

The main components of this dataflow analysis are described below:

1. Main algorithm: iterative worklist with reverse post-order traversal of CFG blocks.

2. Gen Set: for every block `B` of kind `if`-condition, if the condition is a comparison or logical AND of multiple comparisons, it is recorded in Gen set of block `B`.

3. Kill Set: first, all the variables that are defined in block `B` are collected. If any comparison `c` in the function contains any of those variables, `c` must be added to the Kill set of block `B`.
Note that the analysis must be conservative. Since we do not perform pointer analysis, if a comparison contains a pointer dereference, the comparison should be killed at any potential pointer assignment.

4. In Set: for every block `B`, the in set is the intersection of the Out sets of its predecessors.

5. Out Set: for every block `B`, the out set is produced by first subtracting the Kill set of `B`
from the In set. The out set of `B` will be the union of this value and the Gen set of block `B`.

Note that Kill and Gen sets are computed only once for every block. In and Out sets are updated every iteration until Out sets reach a fixed point.

After the dataflow analysis is done, Kill and In sets will be used in proving bounds declarations. A fact is true at a point in program if it is a member of the In set, and not a member of the Kill set.