Implemented the "*" operator for _MMSafe_ptr. (#667)
When a mmsafe_ptr is dereferenced by the "*" operator and after the
null pointer checking is done, the compiler inserts a
DynamicStructIDCheck() function to check if the ID of the mmsafe_ptr
matches the ID of the pointee (a struct).

The algorithm is straightforward:

1. Extract the ID from the mmsafe_ptr.
2. Extract the ID from the struct (always the first field)
3. Insert an integer comparison instruction (ICmpInst) to compare the
   two IDs.
4. Continue executing if the IDs match; otherwise abort the program.

We also implemented a customized memory allocator and a deallocator,
and added tests for this commit. The customized allocator/deallocator
and the tests are in the checkedc repo in the safe-memory-management branch.