Correctly determine bounds for predefined literals (#673)
Correctly determine bounds for predefined literals.  Section 6.4.2.2 of the C11 Standard defines predefined identifiers: the special identifier __func__ evaluates to a string that is the name of the current function.

In clang, predefined identifiers are represented using the PredefinedExpr class, which wraps a string literal.   The PredefinedExpr class evaluates to an lvalue that is the address of the string literal.   To represent bounds information, follow the same pattern that we use for string literals:.  Bind the result of the PredefinedExpr class to a temporary and use the temporary in the bounds of the expression.   We considered having the class wrap a temporary expression, but that causes existing tests of predefined literals to break and requires more changes to the compiler logic for predefined expressions.