Fix Windows-specific test failures.
Currently Azure Devops clones the top-level repo for checkedc-clang.
We then run automation from this repo.  That automation clones the
compiler again (we plan to move the automation to another repo
so that this less confusing).

The automated testing currently isn't using the nested repo.  It
is running tests from the top-level repo.  This breaks on Windows
because the top-level clone isn't preserving line endings.  There
are certain tests for clang that require UNIX-style line endings,
and break with Windows-style line endings.