Fix an increase in compile time from a refactoring change.
The refactoring of bounds declaration checking for binary operator expressions
appears to have caused a significant increase in compile time.  One change
that was made to always the rhs bounds and lhs rvalue bounds, even if those
bounds are ultimately not need.  Defer the computations of those bounds to
only when they are needed, which matches previous behavior.