Fix an increase in compile time from a refactoring change. (#743)
The refactoring of bounds declaration checking for binary operator expressions appears to have caused a significant increase in compile time.  One change that was made was to always compute the rhs bounds and lhs rvalue bounds, even if those bounds are ultimately not needed.  Defer the computations of those bounds to only when they are needed, which matches previous behavior.