Update directory location in automation for code signing.
When we moved to a monorepo, we adjusted the location of automation
scripts to be clang\automation instead of automation.  We missed
updating a path for sync'ing a repo that contain a scripts for code signing.
This worked on existing ADO pipelines because the old copy of the repo was still
there.  The sync'ing did not have the intended effect, but the operations did
not fail.  When I cloned an ADO pipeline for some experimentation, the
sync'ing failed because there is no old copy of the repo.