Pass FunctionDecl as a parameter to BoundsAnalysis functions
Currently, FunctionDecl (FD) is a member of BoundsAnalysis class. As a result,
we need to pass FD while instantiating a BoundsAnalysis object. We plan to add
a BoundsAnalysis object to CheckBoundsDeclaration class but we do not have a
FunctionDecl in the CheckBoundsDeclaration constructor. So we would not be able
to add the BoundsAnalysis object as long as FD is a member of BoundsAnalysis
class. So we need to decouple FD from BoundsAnalysis. We would pass FD only to
methods which need FD.