Change Kill set to be a mapping between Stmt and Vars whose bounds are killed by the Stmt [3/n]
Currently, the Kill set is a set of variables killed in a block. But we have no
info on "where" exactly a variable is killed inside a block. As a result, for a
given Stmt we cannot determine if the bounds of a variable are valid after the
Stmt or not. So we change the Kill set to be a mapping between the Stmt and the
variables killed by that Stmt inside a particular block.

This info would be used by bounds inference/checking to check the widened
bounds after a given Stmt.